Move momento to behavioral patterns
