+++ src/patterns/behavioral/index.js
@@ -1,9 +1,11 @@
 import ChainOfResponsibility from './chain-of-responsibility';
 import Commander from './command';
 import PubSub from './pubsub';
+import Momento from './momento';
 
 module.exports = {
 	ChainOfResponsibility,
 	Commander,
+	Momento,
 	PubSub,
 };
