+++ src/patterns/behavioral/momento.js
