+++ src/patterns/structural/index.js
@@ -1,7 +1,5 @@
 import bridge from './bridge';
-import Momento from './momento';
 
 module.exports = {
 	bridge,
-	Momento,
 };
